+++ src/main/java/org/jboss/ci/tracker/server/CategorizationServiceImpl.java
@@ -1,4 +1,4 @@
-/* 
+/*
  * Copyright (C) 2015 Jan Tymel
  *
  * This program is free software: you can redistribute it and/or modify
@@ -37,7 +37,7 @@ public class CategorizationServiceImpl extends RemoteServiceServlet implements C
     }
 
     @Override
-    public Long saveCategorization(CategorizationDto categorization) {
+    public Integer saveCategorization(CategorizationDto categorization) {
         return categorizationServiceBean.saveCategorization(categorization);
     }
